feat: add standalone agent
Adds a new agent module which wraps EZE into a standalone command line
executable. The new module comes with a docker profile which will
produce a Docker image that will expose the 26500 port by default and
can be used to run EZE standalone.

Tests are minimal - testing the Docker image would require an additional
integration/QA module as the build pipeline relies on Jib and not on a
standard Dockerfile, which means we can't use Testcontainers to build
the agent on the fly in the tests.