build: push agent Docker image on release
Adds a new step to the release pipeline to push the Docker image. This
is triggered only on releases and will not push a SNAPSHOT image.

Keep in mind that it will use the code in the repo as is, so technically
the MANIFEST would still contain the wrong version of the executable,
but the Docker image tag will be correctly overridden with the release
tag.

This is untested, and I'm happy to drop this commit for the initial PR.